+++ R/long2wide.R
@@ -1,6 +1,4 @@
-#' Long to wide format
-#' 
-#' Converts data from long format to wide
+#' Convert data from long to wide format
 #' 
 #' @param .data the tbl in long format
 #' @param within the names of the within column(s)
@@ -12,11 +10,21 @@
 #' 
 #' @examples 
 #' df_long <- sim_design(2, 2, long = TRUE)
-#' long2wide(df_long, "A", "B", "val", "sub_id")
+#' long2wide(df_long, "A", "B")
 #' 
 #' @export
 #' 
-long2wide <- function(.data, within = c(), between = c(), dv = "val", id = "sub_id") {
+long2wide <- function(.data, within = c(), between = c(), dv = "y", id = "id") {
+  if ("design" %in% names(attributes(.data))) {
+    # get parameters from design
+    design <- attributes(.data)$design
+    
+    within <- names(design$within)
+    between <- names(design$between)
+    dv <- design$dv
+    id = design$id
+  }
+  
   d1 <- dplyr::select(.data , tidyselect::one_of(c(id, between, within, dv))) 
   if (length(within)) {
     d1 <- tidyr::unite(d1, ".tmpwithin.", tidyselect::one_of(within))
@@ -31,15 +39,14 @@ long2wide <- function(.data, within = c(), between = c(), dv = "val", id = "sub_
   dplyr::ungroup(d1)
 }
 
-#' Wide to long format
-#' 
-#' Converts data from wide format to long
+#' Convert data from wide to long format
 #' 
 #' @param .data the tbl in wide format
 #' @param within_factors the names of the within factors
 #' @param within_cols the names (or indices) of the within-subject (value) columns
-#' @param dv the name of the dv column (defaults to "val")
-#' @param sep Separator for within-columns (see tidyr::separate)
+#' @param dv the name of the dv column (defaults to "y")
+#' @param id the name of the ID column(s) if they don't exist, a new column will be made
+#' @param sep separator for within-columns (see tidyr::separate)
 #' 
 #' @return a tbl in long format
 #' 
@@ -48,7 +55,19 @@ long2wide <- function(.data, within = c(), between = c(), dv = "val", id = "sub_
 #' 
 #' @export
 #' 
-wide2long <- function(.data, within_factors = c(), within_cols = c(), dv = "val", id = NULL, sep = "[^[:alnum:]]+") {
+wide2long <- function(.data, within_factors = c(), within_cols = c(), 
+                      dv = "y", id = NULL, sep = "[^[:alnum:]]+") {
+  if ("design" %in% names(attributes(.data))) {
+    # get parameters from design
+    design <- attributes(.data)$design
+    
+    dv <- design$dv
+    id = design$id
+    within_factors <- names(design$within)
+    within_cols <- cell_combos(design$within, dv) 
+  }
+  
+  
   if (is.numeric(within_cols)) {
     within_cols <- names(.data)[within_cols]
   }