+++ tests/testthat/test-get_design_long.R
@@ -11,8 +11,6 @@ test_that("2w", {
   
   expect_equal(d$within, d2$within)
   expect_equivalent(d$between, d2$between)
-  expect_equal(d$cells_w, d2$cells_w) 
-  expect_equal(d$cells_b, d2$cells_b) 
   expect_equal(d$n, d2$n)
   expect_equal(d$mu, d2$mu)
   expect_equal(d$sd, d2$sd)
@@ -22,7 +20,7 @@ test_that("2w", {
 # get_design_long ----
 test_that("get_design_long", {
   design <- check_design(2, 2, n = 10, mu = 5, sd = 2, r = 0.5)
-  df_long <- sim_design_(design, long = TRUE, empirical = TRUE)
+  df_long <- sim_design(design = design, long = TRUE, empirical = TRUE)
   d <- get_design_long(df_long, plot = FALSE)
   
   n <- data.frame(A1 = c(10, 10), A2 = c(10, 10), row.names = c("B1", "B2"))
@@ -41,9 +39,9 @@ test_that("get_design_long", {
   df_long <- sim_design(c(2, 2, 2), c(2, 2, 2), long = TRUE, empirical = TRUE)
   d <- get_design_long(df_long, plot = FALSE)
   
-  expect_equal(d$cells_w, c("A1_B1_C1", "A2_B1_C1", "A1_B2_C1", "A2_B2_C1", 
+  expect_equal(d$mu %>% names(), c("A1_B1_C1", "A2_B1_C1", "A1_B2_C1", "A2_B2_C1", 
                             "A1_B1_C2", "A2_B1_C2", "A1_B2_C2", "A2_B2_C2"))
-  expect_equal(d$cells_b, c("D1_E1_F1", "D2_E1_F1", "D1_E2_F1", "D2_E2_F1", 
+  expect_equal(d$mu %>% row.names(), c("D1_E1_F1", "D2_E1_F1", "D1_E2_F1", "D2_E2_F1", 
                             "D1_E1_F2", "D2_E1_F2", "D1_E2_F2", "D2_E2_F2"))
   expect_equal(d$n  %>% sum(), 6400)
   expect_equal(d$mu %>% sum(), 0)
@@ -63,8 +61,6 @@ test_that("2w*2b", {
   d2 <- get_design_long(data, plot = FALSE)
   expect_equal(d$within, d2$within)
   expect_equivalent(d$between, d2$between)
-  expect_equal(d$cells_w, d2$cells_w) 
-  expect_equal(d$cells_b, d2$cells_b) 
   expect_equal(d$n, d2$n)
   expect_equal(d$mu, d2$mu)
   expect_equal(d$sd, d2$sd)