+++ tests/testthat/test-interactive_design.R
@@ -15,14 +15,16 @@ test_that("test", {
     B1 = .5,
     B2 = c(.4, .5, .6)
   )
-  des2 <- check_design(within, between, n, mu, sd, r, plot = FALSE)
+  des2 <- check_design(within, between, n, mu, sd, r, plot = TRUE)
   
   des <- interactive_design()
   
   expect_equal(des$within,  des2$within)
   expect_equal(des$between, des2$between)
-  expect_equal(des$cell_n,  des2$cell_n)
-  expect_equal(des$cell_mu, des2$cell_mu)
-  expect_equal(des$cell_sd, des2$cell_sd)
-  expect_equal(des$cell_r,  des2$cell_r)
+  expect_equal(des$id, des2$id)
+  expect_equal(des$dv, des2$dv)
+  expect_equal(des$n,  des2$n)
+  expect_equal(des$mu, des2$mu)
+  expect_equal(des$sd, des2$sd)
+  expect_equal(des$r,  des2$r)
 })