+++ tests/testthat/test-long2wide.R
@@ -84,3 +84,21 @@ test_that("wide2long", {
   expect_equal(nrow(long_iris), 600)
   expect_equal(names(long_iris), c("Species", "id", "feature", "dimension", "value"))
 })
+
+# from design ----
+test_that("from design", {
+  # w <- sim_design(c(2,2), c(2,2))
+  # w2 <- wide2long(w)
+  # w3 <- long2wide(w2)
+  # expect_equal(names(w2), c("sub_id", "C", "D", "A", "B","rt"))
+  # expect_equal(w, w3)
+  # 
+  # l <- sim_design(c(2,2), c(2,2), long = TRUE)
+  # l2 <- long2wide(l)
+  # expect_equal(names(l2), c("id", "C", "D", "A1_B1", "A1_B2", "A2_B1", "A2_B2"))
+  # 
+  # # from data not made by faux
+  # data <- fr4 %>% 
+  #   dplyr::group_by(rater_id, rater_sex, face_eth) %>% 
+  #   dplyr::summarise(rating = mean(rating))
+})